Refactored deep links controller and registered service
