+++ simplified-accounts-registry/src/main/java/org/nypl/simplified/accounts/registry/DeepLinksControllerType.kt
@@ -1,0 +1,7 @@
+package org.nypl.simplified.accounts.registry
+
+interface DeepLinksControllerType {
+
+
+
+}
