--- simplified-main/src/main/java/org/nypl/simplified/main/DeepLinksControllerType.kt
@@ -1,7 +1,0 @@
-package org.nypl.simplified.profiles.controller.api
-
-interface DeepLinksControllerType {
-
-
-
-}
